- Renamed 'crediter' to 'creditor' wherever I could find it. It's now meant to be understood as as definition (2) in http://www.dictionary.com/browse/creditor. - Also made a couple of small tweaks to cloudformation.yaml to properly allow for multi-stack use, similar to: https://github.com/guardian/zuora-auto-cancel/pull/8
